+++ covasim/datasets/data_loader.py
@@ -2,6 +2,7 @@ import requests
 import os
 import json
 dirname = os.path.dirname(__file__)
+mapper_file_path = os.path.join(dirname,"sources/lookup.json")
 
 
 class DataLoader:
@@ -14,7 +15,6 @@ class DataLoader:
         translated = self.translate(data_json)
         with open(self.file_path(), "a") as outfile:
             json.dump(translated, outfile)
-        self.data = data_json
 
     def fetch_data(self):
         r = requests.get(self.URL)
@@ -36,3 +36,43 @@ class DataLoader:
 
     def translate(self, json):
         raise NotImplementedError
+
+class NeherLabPop(DataLoader):
+    URL="https://raw.githubusercontent.com/neherlab/covid19_scenarios/master/src/assets/data/country_age_distribution.json"
+    FILENAME="neherlab"
+
+    def countries(self):
+        self.load_data()
+        return self.data.keys()
+
+    def data_for_country(self, country):
+        self.load_data()
+        return self.data[country]
+
+    def translate(self, json=False):
+        result = {}
+        for location in json:
+            country = location["country"]
+            age_distribution = location["ageDistribution"]
+            total_pop = sum(age_distribution.values())
+            local_pop = []
+
+            for age, age_pop in age_distribution.items():
+                if age[-1] == '+':
+                    val = [int(age[:-1]), 130, age_pop/total_pop]
+                else:
+                    ages = age.split('-')
+                    val = [int(ages[0]), int(ages[1]), age_pop/total_pop]
+                local_pop.append(val)
+            result[country] = local_pop
+
+        return result
+
+
+def load_country_pop(country):
+    with open(mapper_file_path) as f:
+        mapper = json.load(f)
+    mapper = json.loads(mapper)
+    data_loader = eval("{0}".format(mapper[country]))()
+    return data_loader.data_for_country(country)
+
