+++ covasim/datasets/update_sources.py
@@ -1,21 +1,16 @@
 import requests
 import os
 import json
-import covasim.datasets.translators as t
+from covasim.datasets.data_loader import load_country_pop
+from covasim.datasets.data_loader import NeherLabPop
 dirname = os.path.dirname(__file__)
 mapper_file_path = os.path.join(dirname,"sources/lookup.json")
 
 
-def load_country_pop(country):
-    with open(mapper_file_path) as f:
-        mapper = json.load(f)
-    mapper = json.loads(mapper)
-    data_loader = eval("t.{0}".format(mapper[country]))()
-    return data_loader.data_for_country(country)
 
 
 sources = [
-        t.NeherLabPop(),
+    NeherLabPop(),
 ]
 
 def load_sources():
