+++ tests/datasets/mock_datasets/population/neher_labs.json
@@ -1,6 +1,4 @@
-{
-  "Afghanistan": {
-    "0-9": 100,
+[{"country": "Afghanistan", "ageDistribution": {"0-9": 100,
     "10-19": 200,
     "20-29": 300,
     "30-39": 450,
@@ -8,17 +6,4 @@
     "50-59": 100,
     "60-69": 200,
     "70-79": 30,
-    "80+": 10
-  },
-  "Albania": {
-    "0-9": 333832,
-    "10-19": 361777,
-    "20-29": 472678,
-    "30-39": 390771,
-    "40-49": 323020,
-    "50-59": 386091,
-    "60-69": 330244,
-    "70-79": 194668,
-    "80+": 84716
-  }
-}
+    "80+": 10}}, {"country": "Albania", "ageDistribution": {"0-9": 333832, "10-19": 361777, "20-29": 472678, "30-39": 390771, "40-49": 323020, "50-59": 386091, "60-69": 330244, "70-79": 194668, "80+": 84716}}]
