+++ tests/datasets/mock_datasets/population/neher_labs_translated.json
@@ -1,0 +1,96 @@
+{
+   "Afghanistan":[
+      [
+         0,
+         9,
+         0.2848497926125552
+      ],
+      [
+         10,
+         19,
+         0.25229844731390455
+      ],
+      [
+         20,
+         29,
+         0.18073902191321004
+      ],
+      [
+         30,
+         39,
+         0.11648699681427505
+      ],
+      [
+         40,
+         49,
+         0.0761260207696988
+      ],
+      [
+         50,
+         59,
+         0.04727144192746956
+      ],
+      [
+         60,
+         69,
+         0.027165218557326287
+      ],
+      [
+         70,
+         79,
+         0.012342035090369241
+      ],
+      [
+         80,
+         130,
+         0.002721025001191291
+      ]
+   ],
+   "Albania":[
+      [
+         0,
+         9,
+         0.11600262283962351
+      ],
+      [
+         10,
+         19,
+         0.125713175738247
+      ],
+      [
+         20,
+         29,
+         0.16424994535750784
+      ],
+      [
+         30,
+         39,
+         0.13578824357659697
+      ],
+      [
+         40,
+         49,
+         0.11224558229784797
+      ],
+      [
+         50,
+         59,
+         0.13416199961289835
+      ],
+      [
+         60,
+         69,
+         0.1147558358007879
+      ],
+      [
+         70,
+         79,
+         0.06764479912933401
+      ],
+      [
+         80,
+         130,
+         0.029437795647156487
+      ]
+   ]
+}
